Move `appStore` into `web-common/layout`
