+++ web-common/src/features/models/workspace/ModelWorkspace.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { EntityType } from "@rilldata/web-common/features/entity-management/types";
-  import { appStore } from "@rilldata/web-local/lib/application-state-stores/app-store";
+  import { appStore } from "@rilldata/web-common/layout/app-store";
   import { WorkspaceContainer } from "../../../layout/workspace";
   import ModelInspector from "./inspector/ModelInspector.svelte";
   import ModelBody from "./ModelBody.svelte";
